Fix NDArray check when wrapping observations
This would previously break the DictObsActionWrapper when using it with
an environment that does not natively come with dictionary observations.

Since the type of `observation` would typically by `numpy.ndarray`,
which is not equal to `nptyping.NDArray`, the assertion would fail. What
we actually want to check here is that `observation` has an interface
that is compatible with `nptyping.NDArray` though.